+++ ext/ffi/callback.rs
@@ -1,3 +1,5 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+
 use crate::symbol::NativeType;
 use crate::{
   check_unstable, FfiPermissions, FfiState, ForeignFunction,
